Generate main.py so we can get independent of ipython
