+++ python/bin/BUILD.bazel
@@ -1,4 +1,5 @@
 load("//python/private:interpreter.bzl", _interpreter_binary = "interpreter_binary")
+load("//python/private:repl.bzl", "generate_repl_main")
 load("//python:py_binary.bzl", "py_binary")
 
 filegroup(
@@ -24,27 +25,38 @@ label_flag(
     build_setting_default = "//python:none",
 )
 
+generate_repl_main(
+    name = "repl_py",
+    out = "repl.py",
+    src = "repl_stub.py",
+)
+
+label_flag(
+    name = "repl_stub",
+    build_setting_default = "repl_stub.py",
+)
+
+# The user can modify this flag to make an interpreter shell library available
+# for the stub. E.g. if they switch the stub for an ipython-based on, then they
+# can point this at their version of ipython.
+label_flag(
+    name = "repl_stub_dep",
+    build_setting_default = "//python/private:empty",
+)
+
 py_binary(
     name = "repl",
-    srcs = ["repl.py"],
+    srcs = [":repl.py"],
     deps = [
+        ":repl_stub_dep",
         ":repl_dep",
-        ":repl_lib_dep",
     ],
     visibility = ["//visibility:public"],
 )
 
-# The user can modify this flag to make arbitrary libraries available for import
-# on the REPL. Anything that exposes PyInfo can be used here.
+# The user can modify this flag to make arbitrary PyInfo targets available for
+# import on the REPL.
 label_flag(
     name = "repl_dep",
-    build_setting_default = "//python:none",
-)
-
-# The user can modify this flag to make additional libraries available
-# specifically for the purpose of interacting with the REPL. For example, point
-# this at ipython in your .bazelrc file.
-label_flag(
-    name = "repl_lib_dep",
-    build_setting_default = "//python:none",
+    build_setting_default = "//python/private:empty",
 )
