+++ python/bin/repl_stub.py
@@ -1,0 +1,2 @@
+import code
+code.interact(local=dict(globals(), **locals()))
